--- .eslintcache
@@ -1,1 +1,0 @@
-[{"/Users/anhdai/Project/Webhouse.FE/src/reportWebVitals.ts":"1","/Users/anhdai/Project/Webhouse.FE/src/index.tsx":"2","/Users/anhdai/Project/Webhouse.FE/src/App.tsx":"3","/Users/anhdai/Project/Webhouse.FE/src/components/Navbar.tsx":"4","/Users/anhdai/Project/Webhouse.FE/src/pages/About.tsx":"5","/Users/anhdai/Project/Webhouse.FE/src/pages/Home.tsx":"6","/Users/anhdai/Project/Webhouse.FE/src/features/counter/index.ts":"7","/Users/anhdai/Project/Webhouse.FE/src/components/counter/Counter.tsx":"8","/Users/anhdai/Project/Webhouse.FE/src/features/counter/actionTypes.ts":"9","/Users/anhdai/Project/Webhouse.FE/src/features/counter/selectors.ts":"10","/Users/anhdai/Project/Webhouse.FE/src/features/counter/counterReducer.ts":"11","/Users/anhdai/Project/Webhouse.FE/src/store.ts":"12","/Users/anhdai/Project/Webhouse.FE/src/components/Footer.tsx":"13","/Users/anhdai/Project/Webhouse.FE/src/components/TemplateItem.tsx":"14","/Users/anhdai/Project/Webhouse.FE/src/pages/Templates.tsx":"15","/Users/anhdai/Project/Webhouse.FE/src/pages/Contact.tsx":"16","/Users/anhdai/Project/Webhouse.FE/src/customize-windmill.ts":"17","/Users/anhdai/Project/Webhouse.FE/src/OnlyHeaderLayout.tsx":"18","/Users/anhdai/Project/Webhouse.FE/src/BasicLayout.tsx":"19","/Users/anhdai/Project/Webhouse.FE/src/pages/DetailTemplate.tsx":"20"},{"size":425,"mtime":1612406142341,"results":"21","hashOfConfig":"22"},{"size":813,"mtime":1613753049610,"results":"23","hashOfConfig":"22"},{"size":1057,"mtime":1613924442960,"results":"24","hashOfConfig":"22"},{"size":6354,"mtime":1614091522384,"results":"25","hashOfConfig":"22"},{"size":686,"mtime":1612487227199,"results":"26","hashOfConfig":"22"},{"size":8355,"mtime":1613925369273,"results":"27","hashOfConfig":"22"},{"size":197,"mtime":1612487227151,"results":"28","hashOfConfig":"22"},{"size":1798,"mtime":1612487227087,"results":"29","hashOfConfig":"22"},{"size":108,"mtime":1612487227103,"results":"30","hashOfConfig":"22"},{"size":129,"mtime":1612487227161,"results":"31","hashOfConfig":"22"},{"size":450,"mtime":1612487227142,"results":"32","hashOfConfig":"22"},{"size":372,"mtime":1612487227222,"results":"33","hashOfConfig":"22"},{"size":4414,"mtime":1613291646562,"results":"34","hashOfConfig":"22"},{"size":450,"mtime":1613816739424,"results":"35","hashOfConfig":"22"},{"size":9684,"mtime":1613817508248,"results":"36","hashOfConfig":"22"},{"size":3894,"mtime":1613752983755,"results":"37","hashOfConfig":"22"},{"size":418,"mtime":1613753049361,"results":"38","hashOfConfig":"22"},{"size":345,"mtime":1613924424558,"results":"39","hashOfConfig":"22"},{"size":383,"mtime":1613924474242,"results":"40","hashOfConfig":"22"},{"size":599,"mtime":1613923637963,"results":"41","hashOfConfig":"22"},{"filePath":"42","messages":"43","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"44"},"17jsrr3",{"filePath":"45","messages":"46","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"47","messages":"48","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"49","messages":"50","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"51","messages":"52","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"53"},{"filePath":"54","messages":"55","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"44"},{"filePath":"56","messages":"57","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"58","messages":"59","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"60","messages":"61","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"62","messages":"63","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"64"},{"filePath":"65","messages":"66","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"67","messages":"68","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"69","messages":"70","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"71"},{"filePath":"72","messages":"73","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"74","messages":"75","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"76"},{"filePath":"77","messages":"78","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"79"},{"filePath":"80","messages":"81","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"76"},{"filePath":"82","messages":"83","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"84","messages":"85","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"76"},{"filePath":"86","messages":"87","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"76"},"/Users/anhdai/Project/Webhouse.FE/src/reportWebVitals.ts",[],["88","89"],"/Users/anhdai/Project/Webhouse.FE/src/index.tsx",[],"/Users/anhdai/Project/Webhouse.FE/src/App.tsx",[],"/Users/anhdai/Project/Webhouse.FE/src/components/Navbar.tsx",[],"/Users/anhdai/Project/Webhouse.FE/src/pages/About.tsx",[],["90","91"],"/Users/anhdai/Project/Webhouse.FE/src/pages/Home.tsx",[],"/Users/anhdai/Project/Webhouse.FE/src/features/counter/index.ts",[],"/Users/anhdai/Project/Webhouse.FE/src/components/counter/Counter.tsx",[],"/Users/anhdai/Project/Webhouse.FE/src/features/counter/actionTypes.ts",[],"/Users/anhdai/Project/Webhouse.FE/src/features/counter/selectors.ts",[],["92","93"],"/Users/anhdai/Project/Webhouse.FE/src/features/counter/counterReducer.ts",[],"/Users/anhdai/Project/Webhouse.FE/src/store.ts",[],"/Users/anhdai/Project/Webhouse.FE/src/components/Footer.tsx",[],["94","95"],"/Users/anhdai/Project/Webhouse.FE/src/components/TemplateItem.tsx",[],"/Users/anhdai/Project/Webhouse.FE/src/pages/Templates.tsx",[],["96","97"],"/Users/anhdai/Project/Webhouse.FE/src/pages/Contact.tsx",[],["98","99"],"/Users/anhdai/Project/Webhouse.FE/src/customize-windmill.ts",[],"/Users/anhdai/Project/Webhouse.FE/src/OnlyHeaderLayout.tsx",[],"/Users/anhdai/Project/Webhouse.FE/src/BasicLayout.tsx",[],"/Users/anhdai/Project/Webhouse.FE/src/pages/DetailTemplate.tsx",[],{"ruleId":"100","replacedBy":"101"},{"ruleId":"102","replacedBy":"103"},{"ruleId":"100","replacedBy":"104"},{"ruleId":"102","replacedBy":"105"},{"ruleId":"100","replacedBy":"106"},{"ruleId":"102","replacedBy":"107"},{"ruleId":"100","replacedBy":"108"},{"ruleId":"102","replacedBy":"109"},{"ruleId":"100","replacedBy":"110"},{"ruleId":"102","replacedBy":"111"},{"ruleId":"100","replacedBy":"112"},{"ruleId":"102","replacedBy":"113"},"no-native-reassign",["114"],"no-negated-in-lhs",["115"],["114"],["115"],["114"],["115"],["114"],["115"],["114"],["115"],["114"],["115"],"no-global-assign","no-unsafe-negation"]
