+++ .gitignore
@@ -4,7 +4,6 @@
 /node_modules
 /.pnp
 .pnp.js
-.eslintcache
 
 # testing
 /coverage
@@ -22,3 +21,6 @@
 npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
+
+# cache
+.eslintcache
