+++ src/components/Navbar.tsx
@@ -1,9 +1,9 @@
-import React, { useEffect, useState } from "react";
+import React, { useState } from "react";
 import { NavLink } from "react-router-dom";
 import { Dropdown, DropdownItem } from "@windmill/react-ui";
 
-import { ContactModal } from "components/ContactModal"
 import { TNavbar } from "./Navbar.type"
+import { ContactPopupItem } from 'components/ContactPopupItem'
 
 import logo from "assets/img/logo.svg";
 import icon_desktop from "assets/img/detail-template/icon-destop_view.png";
@@ -13,19 +13,12 @@ import "./Navbar.css";
 export const Navbar: React.FC<TNavbar> = ({type: navbarType = 'BasicNav'}) => {
   const [active, setActive] = useState(false);
   const [isOpenDropdown, setIsOpenDropdown] = useState(false);
-  const [isModalOpen, setIsModalOpen] = useState(false);
 
   const handleClick = () => {
     setActive(!active);
   };
-  const onUpdateOpenModal = () => {
-    setIsModalOpen(false)
-  };
   const toggleDropdown = () => {
     setIsOpenDropdown(!isOpenDropdown);
-  };
-  const toggleContactModal = () => {
-    setIsModalOpen(!isModalOpen)
   };
   return (
     <>
@@ -111,9 +104,8 @@ export const Navbar: React.FC<TNavbar> = ({type: navbarType = 'BasicNav'}) => {
               </NavLink>
               <div
                 className="lg:inline-flex lg:w-auto w-full px-3 py-2 rounded items-center justify-center hover:text-blue-500 text-base md:text-lg font-light cursor-pointer"
-                onClick={toggleContactModal}
               >
-                Contact
+                <ContactPopupItem/>
               </div>
             </div>
             <div className="lg:inline-flex lg:flex-row lg:w-auto w-full lg:items-center items-start flex flex-col lg:h-auto ml-auto">
@@ -127,7 +119,6 @@ export const Navbar: React.FC<TNavbar> = ({type: navbarType = 'BasicNav'}) => {
           </div>
         </nav>
       }
-      <ContactModal modalOpen={isModalOpen} cbUpdateModal={onUpdateOpenModal}/>
       {'SwitchModeNav' === navbarType&&
         <nav className="container mx-auto grid grid-flow-col grid-cols-3 py-3.5 px-1.5 lg:px-0">
           <NavLink to="/" className="inline-flex items-center">
